Move 'loadFonts' function into separate file to resolve circular dependency
